Use default user-agent when overriding special headers.
Accept-Language is overridden through `setUserAgentOverride`
method in Network domain, which declares User-Agent to be a
required parameter. As we don't always care to set User-Agent
at the same time as we set Accept-Language, we store chrome
own user agent as reported on `/json/version` endpoint, and
reuse it if needed.